+++ AllaganLib.Interface/FormFields/ChoiceFormField.cs
@@ -9,36 +9,37 @@ using ImGuiNET;
 
 namespace AllaganLib.Interface.FormFields;
 
-public abstract class ChoiceFormField<T, TS> : FormField<T, TS>
-    where T : IComparable
-    where TS : IConfigurable<T?>
+public abstract class ChoiceFormField<TValue, TConfiguration> : FormField<TValue, TConfiguration>
+    where TValue : IComparable
+    where TConfiguration : IConfigurable<TValue?>
 {
     public ChoiceFormField(ImGuiService imGuiService)
         : base(imGuiService)
     {
     }
 
-    public override T CurrentValue(TS configurable)
+    public override TValue CurrentValue(TConfiguration configurable)
     {
         return configurable.Get(this.Key) ?? this.DefaultValue;
     }
 
-    public override void UpdateFilterConfiguration(TS configurable, T? newValue)
+    public override void UpdateFilterConfiguration(TConfiguration configurable, TValue? newValue)
     {
         configurable.Set(this.Key, newValue);
     }
 
-    public abstract Dictionary<T, string> Choices { get; }
+    public abstract Dictionary<TValue, string> Choices { get; }
 
-    public virtual string GetFormattedChoice(T choice)
+    public virtual string GetFormattedChoice(TValue choice)
     {
         return this.Choices.SingleOrDefault(c => c.Key.Equals(choice)).Value;
     }
 
-    public override void DrawInput(TS configuration, int? inputSize = null)
+    public override bool DrawInput(TConfiguration configuration, int? inputSize = null)
     {
         var choices = this.Choices;
         var activeChoice = this.CurrentValue(configuration);
+        var wasUpdated = false;
 
         var currentSearchCategory = this.GetFormattedChoice(activeChoice);
         ImGui.SetNextItemWidth(inputSize ?? this.InputSize);
@@ -56,10 +57,17 @@ public abstract class ChoiceFormField<T, TS> : FormField<T, TS>
 
                     if (ImGui.Selectable(text, currentSearchCategory == text))
                     {
-                        this.UpdateFilterConfiguration(configuration, item.Key);
+                        if (this.AutoSave)
+                        {
+                            this.UpdateFilterConfiguration(configuration, item.Key);
+                        }
+
+                        wasUpdated = true;
                     }
                 }
             }
         }
+
+        return wasUpdated;
     }
 }