+++ AllaganLib.Interface/FormFields/TimeSpanFormField.cs
@@ -20,23 +20,34 @@ public abstract class TimeSpanFormField<T> : FormField<(TimeUnit, int)?, T>
         return configuration.Get(this.Key) ?? this.DefaultValue;
     }
 
-    public override void DrawInput(T configuration, int? inputSize = null)
+    public override bool DrawInput(T configuration, int? inputSize = null)
     {
         var currentValue = this.CurrentValue(configuration);
         var timeUnit = currentValue?.Item1 ?? null;
         var timeValue = currentValue?.Item2 ?? null;
+        var wasUpdated = false;
 
         if (this.timeSpanPickerWidget.Draw(this.Key + "Input", ref timeUnit, ref timeValue, inputSize))
         {
             if (timeUnit == null || timeValue == null)
             {
-                this.UpdateFilterConfiguration(configuration, null);
+                if (this.AutoSave)
+                {
+                    this.UpdateFilterConfiguration(configuration, null);
+                }
             }
             else
             {
-                this.UpdateFilterConfiguration(configuration, (timeUnit.Value, timeValue.Value));
+                if (this.AutoSave)
+                {
+                    this.UpdateFilterConfiguration(configuration, (timeUnit.Value, timeValue.Value));
+                }
             }
+
+            wasUpdated = true;
         }
+
+        return wasUpdated;
     }
 
     public override void UpdateFilterConfiguration(T configuration, (TimeUnit, int)? newValue)