+++ CHANGELOG.md
@@ -13,6 +13,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.1.12] - 2024-12-11
+
+### Changes
+
+- AllaganLib.GameData
+  - Class jobs in ClassJobCategory were not calculated properly
+
+- AllaganLib.Interface
+  - All fields now have a AutoSave property
+  - FormField's DrawInput now return a boolean indicating if the value changed like imgui does
+  - The MultiChoice form field now renders a preview and shows which items are selected
+  - Added FloatRangeFormField and UintRangeFormFIeld
+
 ## [1.1.11] - 2024-12-04
 
 ### Changes